Sync HealthCenter agent src: 	Agent 	85631: Decouple OS data source from JMXSourceManager - Fixed timestamp issue that caused data to show on 0:00 	89932: Fix up compile issues on Node 0.12 - Wrap V8 interface usage with Nan calls 	89834: hcmqtt library is missing in JDK 	89862: Carry out dynamic linkage from Node.js API to API connector plugin as part of init

+++ runtime.tools/src/ibmras/monitoring/plugins/j9/cpu/CpuPlugin.cpp
@@ -251,7 +251,7 @@ monitordata* CpuPlugin::pullInt() {
 	return data;
 }
 
-char* CpuPlugin::buildCpuDataLine(jdouble tstamp, jdouble processCPULoad, jdouble systemCPULoad) {
+char* CpuPlugin::buildCpuDataLine(jlong tstamp, jdouble processCPULoad, jdouble systemCPULoad) {
 	std::string sdata;
 	sdata.append("startCPU");
 	sdata.append(DELIMITER);
